+++ extension/package.json
@@ -4,6 +4,7 @@
   "description": "YTCR: YouTube Channel Points",
   "private": true,
   "scripts": {
+    "tailwind": "npx postcss styles.css -o public/tailwind.css --watch",
     "watch": "webpack --mode=development --watch --config config/webpack.config.dev.js",
     "build": "webpack --mode=production --config config/webpack.config.js"
   },
@@ -12,12 +13,15 @@
     "css-loader": "^4.3.0",
     "file-loader": "^6.2.0",
     "mini-css-extract-plugin": "^0.10.1",
+    "postcss-cli": "^10.1.0",
     "size-plugin": "^2.0.2",
+    "tailwindcss": "^3.3.2",
     "webpack": "^5.72.0",
     "webpack-cli": "^3.3.12",
     "webpack-merge": "^5.8.0"
   },
   "dependencies": {
+    "autoprefixer": "^10.4.14",
     "webpack-build-notifier": "^2.3.0",
     "ws": "^8.5.0"
   }
